// src/models/response/OpportunityCustomFindById.tsx
// Generated by https://quicktype.io

export interface OpportunityCustomFindById {
    alertaVigenciaContrato:    string;
    arquivo:                   string;
    ata:                       string;
    clienteId:                 number;
    cnpjCliente:               string;
    cnpjOrgao:                 string;
    cnpjOrgaoCondutor:         string;
    compulsorio:               string;
    condicaoPagamento:         string;
    dataAdjudicacao:           Date;
    dataAssinatura:            Date;
    dataCadastro:              Date;
    dataCaptacao:              string;
    dataCertame:               string;
    dataCertameInicial:        string;
    dataEnvioDocumentacao:     Date;
    dataEsclarecimento:        Date;
    dataFimContrato:           Date;
    dataImpugnacao:            Date;
    dataInicioContrato:        Date;
    dataProposta:              Date;
    descricaoModalidade:       string;
    descricaoOportunidade:     string;
    diasParaVencimento:        number;
    edital:                    string;
    enderecoEntrega:           string;
    enderecoEnvio:             string;
    horaCertame:               string;
    horaCertameInicial:        string;
    horaProposta:              string;
    linkContrato:              string;
    linkProposta:              string;
    linkTermoReferencia:       string;
    localidade:                string;
    meEpp:                     string;
    modalidade:                string;
    nomeFantasiaCliente:       string;
    nomeFantasiaOrgao:         string;
    nomeFantasiaOrgaoCondutor: string;
    oportunidadeGanha:         string;
    oportunidadeId:            number;
    orgaoCondutorId:           number;
    orgaoContrato:             string;
    orgaoId:                   number;
    pendencia:                 string;
    plataforma:                string;
    prazoEntrega:              number;
    prazoPagamento:            number;
    pregoeiroEmail:            string;
    pregoeiroNome:             string;
    pregoeiroTelefone:         string;
    razaoSocialCliente:        string;
    razaoSocialOrgao:          string;
    razaoSocialOrgaoCondutor:  string;
    registroPreco:             string;
    saldoOportunidade:         number;
    statusFaturamento:         string;
    statusOportunidade:        string;
    tipoArqAta:                string;
    tipoArquivo:               string;
    tipoDisputa:               string;
    uasg:                      string;
    valorFaturado:             number;
    valorTotal:                number;
    valorTotalRecebido:        number;
}